Add a -[ILCrashReporter submitManualCrashReport] method.
Also add a -[ILCrashReporter launchReporter] method, which enables two-stage setup of the ILCrashReporter object: first one initialises the object and can set properties on it, and after the properties are set, use -launchReporter.  This makes it much easier to set several properties on the crash reporter, rather than trying to use the myriad of class methods to set and launch the crash reporter immediately.  (The class +launchReporter... methods have been kept for backwards-compatibility.)